backend/requirements.txt
fastapi
uvicorn
pydantic
numpy
pillow
opencv-python
python-multipart
torch
